% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{as_character}
\alias{as_character}
\title{Convert one dimension data to character}
\usage{
as_character(x)
}
\arguments{
\item{x}{date, numeric, integer, factor. data to be changed.}
}
\value{
character
}
\description{
Convert one dimension data to character
}
\examples{
as_character(pi)
as_character(date())

}
