% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{as_date}
\alias{as_date}
\title{Convert year-month-day format factor/character to date}
\usage{
as_date(x)
}
\arguments{
\item{x}{character or factor. data to be changed.}
}
\value{
Date class
}
\description{
Convert year-month-day format factor/character to date
}
\examples{
as_date("2004-4-11")

}
