% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{as_factor}
\alias{as_factor}
\title{Convert one dimension data to factor}
\usage{
as_factor(x, ordered = FALSE)
}
\arguments{
\item{x}{date, numeric, integer. data to be changed.}

\item{ordered}{logical. Whether the factor to be generated is 
an ordered factor.}
}
\value{
factor
}
\description{
Convert one dimension data to factor
}
\examples{
as_factor(letters)

}
