% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_type.R
\name{as_integer}
\alias{as_integer}
\title{Convert one dimension data to integer}
\usage{
as_integer(x)
}
\arguments{
\item{x}{factor, character. data to be changed.}
}
\value{
integer
}
\description{
Convert one dimension data to integer
}
\examples{
as_integer(pi)
as_integer(letters)

}
