% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_type.R
\name{as_numeric}
\alias{as_numeric}
\title{Convert one dimension data to numeric}
\usage{
as_numeric(x)
}
\arguments{
\item{x}{integer, factor, character. data to be changed.}
}
\value{
numeric
}
\description{
Convert one dimension data to numeric
}
\examples{
as_numeric(3L)
as_numeric("3.14")

}
