% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BitStat.R
\docType{package}
\name{BitStat-package}
\alias{BitStat}
\alias{BitStat-package}
\title{BitStat: Statistical Analytics using R and Shiny}
\description{
BitStat support data diagnosis, exploration, and transformation and data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r2bit/BitStat/}
  \item Report bugs at \url{https://github.com/r2bit/BitStat/issues}
}

}
\author{
\strong{Maintainer}: Choonghyun Ryu \email{choonghyun.ryu@gmail.com}

}
