% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Launch BitStat apps}
\usage{
launch(language = c("kr", "en")[1], ...)
}
\arguments{
\item{language}{character. app language. support Korean and English.}

\item{...}{additional arguments to pass to shiny::runApp (e.g, port = 8080)}
}
\description{
Launch BitStat apps
}
\details{
See \url{https://r2bit.github.io/BitStat/docs} for radiant documentation and tutorials
}
\examples{
\dontrun{
launch()
}

}
